MarkAssignments.cpp: fix HasInterferingSideEffects
Before this commit, every time we decided to serialize an Instruction we
were considering other instructions that might have interfering side
effects with it.
This was resulting in many more instructions than necessary being marked
with HasInterferingSideEffects.

This commits fixes this problem, by only checking interfering side
effects when an Instruction is marked as having side effects.