MarkAssignments: drop check on many uses
This commit changes MarkAssignments (and related tests) so that now
instruction with more than one use are not assigned to a new variable
by default.

In some cases it might still be good to serialize them for readability,
but for those cases we will implementa a standalone pass that only
implements heuristics based on readability.

MarkAssignments now only takes decisions based on semantics of
side-effects.