MarkAssignments: emit LocalVariable for SelectInst
This is a temporary workaround for avoiding strings of exponential size
in the decompiled C code.
It is known not to be very robust, in that it's not guaranteed that it
will completely avoid the problem in the future, but it should cover the
most part of common problems.

A better fix would be to have a dedicated pass that looks at the
dataflow graph and decided where are the best points to inject local
variables, instead of blindly deciding to emit local variables for
SelectInst.
However, this would require much more work, and this problem is already
stalling merging features in revng, so for now we adopted this.