BeautifyGHAST: reorder `simplifyImplicitContinue`
Move the `simplifyImplicitContinue` at the end of the beautify pass,
after all the pass that modify the anatomy of the ASTTree.

Before this change, the pass was run before some changes that moved some
`continue`s in positions where they are not considered implicit.

Added an assertion which checks that during the `noFallThrough`
promotion phase, no implicit `ContinueNode` is present.