DLA: don't link formal arguments with constant
The inter-procedural part of the DLA frontend connects actual arguments
of function calls with formal arguments of the callee functions, with
instance links at offset 0.
This represents the information that the type pointed-to by the actual
argument has an instance of the type pointed-to by the formal argument
at offset 0.

Before this commit, this was done even when the actual argument was an
integer constant leading to various problems:
1. Most of the times, small integer constants passed as actual
   cannot represent valid pointers, leading to graphs bigger than
   necessary.
2. Even when an integer constant might actually represent a pointer into
   valid memory, if it does it should fall into some segment.
   If it does, by the time DLA runs, those integer constants should have
   already converted into calls to the special SegmentRef opcode, which
   is already handled properly.
3. Finally, those constants end up being very connected in the DLA
   graph, creating connections between other nodes that are otherwise
   unrelated or very far from each other. This pollutes the graph and
   rapidly degrades the quality of the results.

This commit properly guards the code so that the instance-at-offset-0
link between actual and formal arguments is never injected if the actual
argument is an integer constant.