MakeLocalVariables uses initModelTypes
This enables to generate LocalVariables with nicer-looking types,
learning from Stores that write values into them.

RemoveLoadStore had to be updated to handle cstringLiteral as a corner
case.

In the future, we can probably treat cstringLiteral homogeneously if we
change MakeSegmentRefs so to inject AddressOf around cstringLiteral, and
change cstringLiteral itself to have reference semantics.