RegionCFGTreeImpl: do not reprocess conditionals
Do not re-enqueue conditional nodes a for combing, when duplicating a
conditional.

This can be done in such way, thanks to the property that we process
conditional nodes in postorder. Therefore, if conditional node A causes
the duplication of, say, node B, we have the guarantee that such node B
has already been processed by the combing. Therefore, the clone of node
B, say node B', should not be reprocessed for combing.