MakeModelGEP: simplify ModelGEPs LLVM Types
Now ModelGEPs emitted by MakeModelGEP always forward the LLVM Type of
the operand representing the base address.

Before this commit they were occasionally using an integer type with the
size of the underlying `model::QualifiedType`, but that is not necessary
and may end up requiring `llvm::IntegerType`s that are very large, to
the point of not being supported by LLVM, for no additional advantage.